rag-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
psycopg2-binary  # For Neon Postgres - version will be resolved automatically
SQLAlchemy>=2.0.23
asyncpg>=0.29.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
tiktoken>=0.5.2
openai>=1.3.5
python-multipart>=0.0.6
aiofiles>=23.2.1
python-frontmatter>=1.0.0
dashscope>=1.19.2  # Required for Qwen API
transformers>=4.21.0
torch>=1.13.0